pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hardwarelib"
version = "0.1.0"
description = "General-purpose lab hardware driver library"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pyvisa",
    "pyserial",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
include = ["hardwarelib*"]
